my_api/requirements.txt
Django==4.0
psycopg2-binary==2.9.3
djangorestframework==3.14.0
gunicorn==20.1.0
PyYAML==6.0
uritemplate==4.1.1